requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
numpy==2.3.3
pycparser==2.23
pydantic==2.11.10
pydantic-core==2.33.2
pydub==0.25.1
python-dotenv==1.1.1
requests==2.32.5
sarvamai==0.1.21
sniffio==1.3.1
sounddevice==0.5.2
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
websockets==15.0.1
